+++ VAE/layers.h
@@ -12,12 +12,12 @@ public:
 
 	void load_params(std::string filew, std::string fileb); 
 
-	Eigen::MatrixXd W;
-	Eigen::VectorXd b;
-	Eigen::MatrixXd J;
+	Eigen::MatrixXf W;
+	Eigen::VectorXf b;
+	Eigen::MatrixXf J;
 
-	Eigen::VectorXd input;
-	Eigen::VectorXd output;
+	Eigen::VectorXf input;
+	Eigen::VectorXf output;
 
 	void forward();
 	void backward();
@@ -29,19 +29,19 @@ public:
 class LeakyReLU
 {
 public:
-	LeakyReLU(double r)
+	LeakyReLU(float r)
 	{
 		m_r = r; 
 	}
 
-	Eigen::VectorXd input;
-	Eigen::VectorXd output;
-	Eigen::MatrixXd J;
+	Eigen::VectorXf input;
+	Eigen::VectorXf output;
+	Eigen::MatrixXf J;
 
 	void forward();
 	void backward();
 
-	double m_r;
+	float m_r;
 };
 
 
@@ -50,9 +50,9 @@ class Tanh
 public:
 	Tanh() {};
 
-	Eigen::VectorXd input;
-	Eigen::VectorXd output;
-	Eigen::MatrixXd J;
+	Eigen::VectorXf input;
+	Eigen::VectorXf output;
+	Eigen::MatrixXf J;
 
 	void forward();
 	void backward();
@@ -68,26 +68,26 @@ public:
 
 	int jointnum; 
 
-	Eigen::VectorXd input;  // [62 * 6]
-	Eigen::VectorXd output; // [62 * 9]
-	Eigen::MatrixXd J; // [(62*9) * (62*6)] 
+	Eigen::VectorXf input;  // [62 * 6]
+	Eigen::VectorXf output; // [62 * 9]
+	Eigen::MatrixXf J; // [(62*9) * (62*6)] 
 
 	void forward(); 
 	void  backward(); 
 
-	Eigen::VectorXd computey(Eigen::VectorXd x); 
-	Eigen::MatrixXd computeJBlock(Eigen::VectorXd x); 
+	Eigen::VectorXf computey(Eigen::VectorXf x); 
+	Eigen::MatrixXf computeJBlock(Eigen::VectorXf x); 
 
 private:
 	// y = ||x||_2, return dy/dx^T
-	Eigen::Matrix3d computeJVecNorm(Eigen::Vector3d x);
+	Eigen::Matrix3f computeJVecNorm(Eigen::Vector3f x);
 
 	//c = a - (a.dot(b))b (assume b is normalized)
 	// return dc/da^T, dc/db^T
-	void computeJr2(Eigen::Vector3d a, Eigen::Vector3d b, Eigen::Matrix3d& Ja, Eigen::Matrix3d& Jb);
+	void computeJr2(Eigen::Vector3f a, Eigen::Vector3f b, Eigen::Matrix3f& Ja, Eigen::Matrix3f& Jb);
 
 	// c = a.cross(b)
 	// return dc/da^T, dc/db^T
-	void computeJCross(Eigen::Vector3d a, Eigen::Vector3d b, Eigen::Matrix3d d_b_d_a, Eigen::Matrix3d& Ja, Eigen::Matrix3d& Jb);
+	void computeJCross(Eigen::Vector3f a, Eigen::Vector3f b, Eigen::Matrix3f d_b_d_a, Eigen::Matrix3f& Ja, Eigen::Matrix3f& Jb);
 
 };
